Fixed headers for guest user and admin, and created separet component for list
