+++ components/Dropdown.tsx
@@ -1,51 +1,11 @@
 "use client";
-import {getCategories} from "@/app/actions/category";
 import {getProductsByCategory} from "@/app/actions/product";
-import {useEffect, useState} from "react";
-
-interface Category {
-  id: number;
-  name: string;
-}
+import {useState} from "react";
+import {List} from "./List";
 
 interface Product {
   id: number;
   name: string;
-}
-
-export function List({
-  onCategoryClick,
-}: {
-  onCategoryClick: (categoryName: string) => void;
-}) {
-  const [categories, setCategories] = useState<Category[]>([]);
-
-  useEffect(() => {
-    const fetchCategories = async () => {
-      try {
-        const categories = await getCategories();
-        setCategories(categories);
-      } catch (error) {
-        console.error("Error fetching categories:", error);
-      }
-    };
-
-    fetchCategories();
-  }, []);
-
-  return (
-    <>
-      {categories.map(category => (
-        <li
-          className="hover:cursor-pointer text-black py-2 px-5 border-b border-gray-200 hover:bg-gray-200 transition-all"
-          key={category.id}
-          onClick={() => onCategoryClick(category.name)}
-        >
-          {category.name}
-        </li>
-      ))}
-    </>
-  );
 }
 
 export default function Dropdown() {
