+++ components/header/GuestHeader.tsx
@@ -1,9 +1,5 @@
-import Dropdown, { List } from '../Dropdown';
+import Dropdown from "../Dropdown";
 
 export default function GuestHeader() {
-  return (
-    <Dropdown>
-      <List />
-    </Dropdown>
-  );
+  return <Dropdown />;
 }
